refactor: Cookie 따로 폴더 관리하지 않고 Helper Config 클래스로 각각 옮김
