+++ main-server/src/main/kotlin/com/techtaurant/mainserver/security/config/CookieProperties.kt
@@ -1,4 +1,4 @@
-package com.techtaurant.mainserver.security.cookie
+package com.techtaurant.mainserver.security.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
