+++ main-server/src/main/kotlin/com/techtaurant/mainserver/security/helper/CookieHelper.kt
@@ -1,5 +1,6 @@
-package com.techtaurant.mainserver.security.cookie
+package com.techtaurant.mainserver.security.helper
 
+import com.techtaurant.mainserver.security.config.CookieProperties
 import com.techtaurant.mainserver.security.jwt.JwtConstants
 import jakarta.servlet.http.Cookie
 import jakarta.servlet.http.HttpServletRequest
@@ -21,8 +22,8 @@ class CookieHelper(
     }
 
     fun deleteAllAuthCookies(response: HttpServletResponse) {
-        deleteCookie(response, JwtConstants.REFRESH_TOKEN_COOKIE)
-        deleteCookie(response, JwtConstants.ACCESS_TOKEN_COOKIE)
+        deleteCookie(response, JwtConstants.Companion.REFRESH_TOKEN_COOKIE)
+        deleteCookie(response, JwtConstants.Companion.ACCESS_TOKEN_COOKIE)
     }
 
     fun getCookie(request: HttpServletRequest, name: String): String? {
